Fix issues with background task starting , expand filefield of jqfileupload, start test implementation for cases
The jqfileupload files had to be expanded, because sometimes, long filenames lead to
database errors, since the generated filename for the media directories cannot be stored. This
depends on the filename length of the uploaded file, and mostly affects tests, but widening
the field should make the website more robust with regards to longer file names that are uploaded.

Issue #479