Add Philips iSyntax (.isyntax) support via libisyntax (#121)
**Description**
This PR adds support for the proprietary Philips iSyntax whole-slide
image format. Users can now point rse-panimg at .isyntax files and get
back tiled pyramid tiff images. Users do not need to have the Philips
iSyntax SDK installed.

**Changes**
- add `panimg/contrib/wsi_isyntax_to_tiff/isyntax_to_tiff.py` following
the implementation of `panimg/contrib/wsi_dcm_to_tiff/dcm_to_tiff.py` to
convert the iSyntax file to tiff.
- iSyntax specific functions in `panimg/image_builders/tiff.py`
following the handling of DICOM WSIs.
- Add `pyisyntax` as a dependency. It uses the open source `libisyntax`
instead of the Philips iSyntax SDK
- Add isyntax related tests. The integration test
`test_image_builder_tiff` skips the isyntax test file. The test is very
slow as the complete isyntax file is converted. If a smaller image is
desired, I can see if there is a way to get rid of some levels while
keeping the isyntax file format, but as far as I know there is no easy
way. Let me know if the test should be marked as slow so that it can be
skipped in local testing.
- The tests depend on a public isyntax example file (the only one I
could find). It is relatively large to other test images, so it is
downloaded and cached at test time, see `conftest.py`

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>